adicionado ursal.c e ursal.h
